root_agent: split WPF UI from core and categorize tests
