+++ Wrecept.Core.sln
@@ -1,0 +1,28 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.0.31903.59
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Wrecept.Core", "Wrecept.Core\Wrecept.Core.csproj", "{6E56EAF7-F483-402D-841C-60A6AC5748A5}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Wrecept.Core.Tests", "Wrecept.Core.Tests\Wrecept.Core.Tests.csproj", "{3E7FA333-5A11-40E6-879B-FB4FF3ED9F8B}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{6E56EAF7-F483-402D-841C-60A6AC5748A5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6E56EAF7-F483-402D-841C-60A6AC5748A5}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6E56EAF7-F483-402D-841C-60A6AC5748A5}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6E56EAF7-F483-402D-841C-60A6AC5748A5}.Release|Any CPU.Build.0 = Release|Any CPU
+		{3E7FA333-5A11-40E6-879B-FB4FF3ED9F8B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3E7FA333-5A11-40E6-879B-FB4FF3ED9F8B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3E7FA333-5A11-40E6-879B-FB4FF3ED9F8B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3E7FA333-5A11-40E6-879B-FB4FF3ED9F8B}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+EndGlobal
