+++ Wrecept.Core/Models/Invoice.cs
@@ -1,0 +1,11 @@
+using System.Collections.Generic;
+
+namespace Wrecept.Core.Models;
+
+public class Invoice
+{
+    public int Id { get; set; }
+    public int SupplierId { get; set; }
+    public Supplier Supplier { get; set; } = null!;
+    public List<InvoiceItem> Items { get; set; } = new();
+}
