+++ Wrecept.Core/Models/InvoiceItem.cs
@@ -1,0 +1,11 @@
+namespace Wrecept.Core.Models;
+
+public class InvoiceItem
+{
+    public int Id { get; set; }
+    public int InvoiceId { get; set; }
+    public Invoice Invoice { get; set; } = null!;
+    public int ProductId { get; set; }
+    public Product Product { get; set; } = null!;
+    public int Quantity { get; set; }
+}
