+++ Wrecept.Core/Models/Product.cs
@@ -1,0 +1,7 @@
+namespace Wrecept.Core.Models;
+
+public class Product
+{
+    public int Id { get; set; }
+    public string Name { get; set; } = string.Empty;
+}
