+++ Wrecept.Core/Models/Supplier.cs
@@ -1,0 +1,7 @@
+namespace Wrecept.Core.Models;
+
+public class Supplier
+{
+    public int Id { get; set; }
+    public string Name { get; set; } = string.Empty;
+}
