+++ Wrecept.Core/Services/IDemoDataService.cs
@@ -1,0 +1,8 @@
+using System.Threading.Tasks;
+
+namespace Wrecept.Core.Services;
+
+public interface IDemoDataService
+{
+    Task SeedAsync();
+}
