+++ Wrecept.UI/App.xaml
@@ -1,0 +1,6 @@
+<Application x:Class="Wrecept.UI.App"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
+    <Application.Resources>
+    </Application.Resources>
+</Application>
