+++ Wrecept.UI/MainWindow.xaml
@@ -1,0 +1,10 @@
+<Window x:Class="Wrecept.UI.MainWindow"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        xmlns:views="clr-namespace:Wrecept.UI.Views"
+        mc:Ignorable="d"
+        Title="Wrecept" Height="450" Width="800">
+    <views:InvoiceEditorView />
+</Window>
