+++ Wrecept.UI/MainWindow.xaml.cs
@@ -1,0 +1,11 @@
+using System.Windows;
+
+namespace Wrecept.UI;
+
+public partial class MainWindow : Window
+{
+    public MainWindow()
+    {
+        InitializeComponent();
+    }
+}
