+++ Wrecept.UI/Views/InvoiceEditorView.xaml.cs
@@ -1,0 +1,11 @@
+using System.Windows.Controls;
+
+namespace Wrecept.UI.Views;
+
+public partial class InvoiceEditorView : UserControl
+{
+    public InvoiceEditorView()
+    {
+        InitializeComponent();
+    }
+}
