--- Wrecept.WpfApp/App.xaml
@@ -1,6 +1,0 @@
-<Application x:Class="Wrecept.WpfApp.App"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
-    <Application.Resources>
-    </Application.Resources>
-</Application>
