--- Wrecept.WpfApp/MainWindow.xaml.cs
@@ -1,11 +1,0 @@
-using System.Windows;
-
-namespace Wrecept.WpfApp;
-
-public partial class MainWindow : Window
-{
-    public MainWindow()
-    {
-        InitializeComponent();
-    }
-}
