--- Wrecept.WpfApp/Models/InvoiceItem.cs
@@ -1,11 +1,0 @@
-namespace Wrecept.WpfApp.Models;
-
-public class InvoiceItem
-{
-    public int Id { get; set; }
-    public int InvoiceId { get; set; }
-    public Invoice Invoice { get; set; } = null!;
-    public int ProductId { get; set; }
-    public Product Product { get; set; } = null!;
-    public int Quantity { get; set; }
-}
