--- Wrecept.WpfApp/Models/Supplier.cs
@@ -1,7 +1,0 @@
-namespace Wrecept.WpfApp.Models;
-
-public class Supplier
-{
-    public int Id { get; set; }
-    public string Name { get; set; } = string.Empty;
-}
