--- Wrecept.WpfApp/Repositories/IRepository.cs
@@ -1,14 +1,0 @@
-using System.Collections.Generic;
-using System.Threading.Tasks;
-
-namespace Wrecept.WpfApp.Repositories
-{
-    public interface IRepository<T> where T : class
-    {
-        Task<IEnumerable<T>> GetAllAsync();
-        Task<T> GetByIdAsync(int id);
-        Task AddAsync(T entity);
-        Task UpdateAsync(T entity);
-        Task DeleteAsync(T entity);
-    }
-}
