--- Wrecept.WpfApp/Services/IDemoDataService.cs
@@ -1,8 +1,0 @@
-using System.Threading.Tasks;
-
-namespace Wrecept.WpfApp.Services;
-
-public interface IDemoDataService
-{
-    Task SeedAsync();
-}
