--- Wrecept.WpfApp/ViewModels/InvoiceEditorViewModel.cs
@@ -1,12 +1,0 @@
-using System.ComponentModel;
-using System.Runtime.CompilerServices;
-
-namespace Wrecept.WpfApp.ViewModels;
-
-public class InvoiceEditorViewModel : INotifyPropertyChanged
-{
-    public event PropertyChangedEventHandler? PropertyChanged;
-
-    protected void OnPropertyChanged([CallerMemberName] string? propertyName = null)
-        => PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(propertyName));
-}
