--- Wrecept.WpfApp/Views/InvoiceEditorView.xaml
@@ -1,10 +1,0 @@
-<UserControl x:Class="Wrecept.WpfApp.Views.InvoiceEditorView"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-             mc:Ignorable="d">
-    <Grid>
-        <TextBlock Text="Invoice Editor" />
-    </Grid>
-</UserControl>
