--- Wrecept.WpfApp/Views/InvoiceEditorView.xaml.cs
@@ -1,11 +1,0 @@
-using System.Windows.Controls;
-
-namespace Wrecept.WpfApp.Views;
-
-public partial class InvoiceEditorView : UserControl
-{
-    public InvoiceEditorView()
-    {
-        InitializeComponent();
-    }
-}
